refactor(llm): extract pop_sse_line into shared sse module
